F/#187 게시글 검색 정상적으로 수행되도록 하였음.

+++ src/main/java/com/se/apiserver/v1/post/application/service/PostDeleteService.java
@@ -29,7 +29,10 @@ public class PostDeleteService {
                     return attach.getSaveName();
                 })
                 .toArray(String[]::new);
-        multipartFileDeleteService.delete(attachSaveNameList);
+
+        if(attachSaveNameList.length > 0)
+            multipartFileDeleteService.delete(attachSaveNameList);
+
         postJpaRepository.delete(post);
         return true;
     }
